+++ comb_spec_searcher/specification.py
@@ -11,6 +11,7 @@ import sympy
 from logzero import logger
 from sympy import Eq, Expr, Function, Number, solve, var
 
+from comb_spec_searcher.exception import SpecificationNotFound
 from comb_spec_searcher.typing import (
     CombinatorialClassType,
     CombinatorialObjectType,
@@ -24,7 +25,7 @@ from .exception import (
     InvalidOperationError,
     TaylorExpansionError,
 )
-from .isomorphism import AtomEquals, Bijection, Isomorphism
+from .isomorphism import Bijection, Isomorphism
 from .specification_drawer import SpecificationDrawer
 from .strategies import (
     EmptyStrategy,
@@ -58,12 +59,14 @@ class CombinatorialSpecification(
         self,
         root: CombinatorialClassType,
         rules: Iterable[AbstractRule[CombinatorialClassType, CombinatorialObjectType]],
+        group_equiv: bool = True,
     ):
         self.root = root
         self.rules_dict = {rule.comb_class: rule for rule in rules}
         self.labels: Dict[CombinatorialClassType, int] = {}
         self._label_to_tiling: Dict[int, CombinatorialClassType] = {}
-        self._group_equiv_in_path()
+        if group_equiv:
+            self._group_equiv_in_path()
         self._set_subrules()
 
     def _set_subrules(self) -> None:
@@ -132,7 +135,9 @@ class CombinatorialSpecification(
         """
         verification_packs: Dict[CombinatorialClassType, StrategyPack] = {}
         for comb_class, rule in self.rules_dict.items():
-            if isinstance(rule, VerificationRule):
+            if isinstance(rule, VerificationRule) and not isinstance(
+                rule.strategy, EmptyStrategy
+            ):
                 try:
                     verification_packs[comb_class] = rule.pack()
                 except InvalidOperationError:
@@ -144,8 +149,38 @@ class CombinatorialSpecification(
             self._group_equiv_in_path()
             self._set_subrules()
 
+    def unexpanded_verified_classes(self) -> Set[CombinatorialClassType]:
+        """
+        Returns the set of verified classes so you can determine which classes still
+        need to be expanded.
+        """
+        return set(
+            comb_class
+            for comb_class, rule in self.rules_dict.items()
+            if isinstance(rule, VerificationRule)
+            and not isinstance(rule.strategy, EmptyStrategy)
+        )
+
+    def expand_all_verified_with_pack(
+        self, pack: StrategyPack, max_expansion_time: float = 0
+    ) -> None:
+        """
+        Will attempt to expand all verified class with the given strategy pack, limiting
+        expansion to a particular time limit if max_expansion_time is specified.
+        """
+        for comb_class, rule in list(self.rules_dict.items()):
+            if isinstance(rule, VerificationRule) and not isinstance(
+                rule.strategy, EmptyStrategy
+            ):
+                self.expand_comb_class(comb_class, pack, max_expansion_time)
+        self._group_equiv_in_path()
+        self._set_subrules()
+
     def expand_comb_class(
-        self, comb_class: Union[int, CombinatorialClassType], pack: StrategyPack
+        self,
+        comb_class: Union[int, CombinatorialClassType],
+        pack: StrategyPack,
+        max_expansion_time: float = 0,
     ) -> None:
         """
         Will try to expand a particular class with respect to the strategy pack
@@ -155,17 +190,25 @@ class CombinatorialSpecification(
         from .comb_spec_searcher import CombinatorialSpecificationSearcher
 
         rule = self.get_rule(comb_class)
-        self.rules_dict.pop(rule.comb_class)
+
+        removed_rule = self.rules_dict.pop(rule.comb_class)
+
         pack = pack.add_verification(AlreadyVerified(self.rules_dict), apply_first=True)
         css = CombinatorialSpecificationSearcher(rule.comb_class, pack)
         logger.info(css.run_information())
-        rules = css._auto_search_rules()  # pylint: disable=protected-access
-        for rule in rules:
-            if not isinstance(rule.strategy, AlreadyVerified):
-                assert (
-                    rule.comb_class not in self.rules_dict or rule.comb_class.is_empty()
-                )
-                self.rules_dict[rule.comb_class] = rule
+        try:
+            rules = css._auto_search_rules(  # pylint: disable=protected-access
+                max_expansion_time=max_expansion_time
+            )
+            for rule in rules:
+                if not isinstance(rule.strategy, AlreadyVerified):
+                    assert (
+                        rule.comb_class not in self.rules_dict
+                        or rule.comb_class.is_empty()
+                    )
+                    self.rules_dict[rule.comb_class] = rule
+        except SpecificationNotFound:
+            self.rules_dict[rule.comb_class] = removed_rule
 
     def _is_valid_spec(self) -> bool:
         """Checks that each class is on a left hand side."""
@@ -190,7 +233,9 @@ class CombinatorialSpecification(
                     " in the specification."
                 ) from e
         if comb_class not in self.rules_dict:
-            assert comb_class.is_empty(), "rule not in the spec and not empty"
+            assert (
+                comb_class.is_empty()
+            ), f"rule not in the spec and not empty\n{comb_class}"
             empty_strat = EmptyStrategy[
                 CombinatorialClassType, CombinatorialObjectType
             ]()
@@ -398,26 +443,33 @@ class CombinatorialSpecification(
 
         Raise an SanityCheckFailure error if it fails.
         """
-        return self._is_valid_spec() and all(
-            all(rule.sanity_check(n) for rule in self.rules_dict.values())
-            for n in range(length + 1)
-        )
+        if not self._is_valid_spec():
+            return False
+
+        for rule in self.rules_dict.values():
+            try:
+                for n in range(length + 1):
+                    if not rule.sanity_check(n):
+                        return False
+            except NotImplementedError:
+                logger.warning(
+                    "Can't sanity check the rule %s -> %s, which is\n" "%s",
+                    self.get_label(rule.comb_class),
+                    tuple(self.get_label(child) for child in rule.children),
+                    rule,
+                )
+
+        return True
 
     def get_bijection_to(
-        self, other: "CombinatorialSpecification", eq: Optional[AtomEquals] = None
+        self, other: "CombinatorialSpecification"
     ) -> Optional[Bijection]:
         """Get bijection from self to other."""
-        if eq is None:
-            return Bijection.construct(self, other)
-        return Bijection.construct(self, other, eq)
+        return Bijection.construct(self, other)
 
-    def are_isomorphic(
-        self, other: "CombinatorialSpecification", eq: Optional[AtomEquals] = None
-    ) -> bool:
+    def are_isomorphic(self, other: "CombinatorialSpecification") -> bool:
         """Check if self is isomorphic to other."""
-        if eq is None:
-            return Isomorphism.check(self, other)
-        return Isomorphism.check(self, other, eq)
+        return Isomorphism.check(self, other)
 
     def show(
         self, levels_shown: int = 0, levels_expand: int = 0, verbose: bool = False
@@ -461,7 +513,8 @@ class CombinatorialSpecification(
             levels_expand=levels_expand,
             verbose=verbose,
         )
-        sd.share()
+        file_url = sd.share()
+        print(f"The specification is available at {file_url}.")
 
     def __eq__(self, other: object) -> bool:
         if not isinstance(other, CombinatorialSpecification):
@@ -536,7 +589,7 @@ class CombinatorialSpecification(
         """
         root = CombinatorialClass.from_dict(d.pop("root"))
         rules = [AbstractRule.from_dict(rule_dict) for rule_dict in d.pop("rules")]
-        return CombinatorialSpecification(root, rules)
+        return CombinatorialSpecification(root, rules, group_equiv=False)
 
 
 class AlreadyVerified(VerificationStrategy[CombinatorialClass, CombinatorialObject]):